Interface now depends on the World's state
The GameInterface now stores a number of different interfaces for
each world state. When adding an InterfaceElement, a WorldState
must be specified. The InterfaceElement will then only be rendered
and updated when the World is in the required state.

I have assumed that each WorldState will have a different interface.