Made world changing quicker
Separated interface to game interface and world interface. The interface
common to levels (i.e. menu system) is put into game interface. Any
interface elements specific to a world is added to the world interface.
Level changing is a lot quicker as buffers for the menu need not be
reallocated for each world.