Fixed two 'falling through world' bugs.
First one caused Actors to fall through moving platforms. When the Actor was
moved it was not moved enough, causing the Actor and the platform to overlap,
meaning the Actor subsequently ignored that platform when collision checking.

Second caused an Actor to fall through the World by attempting to fall off
a block despite there not being space to fall.